chore: Start implementing support for additional XamlRoot properties
